refactor(core): move DTCG value-to-CSS mappers to core
